API/requirements.txt
numpy
flask
flask_cors
flask_restful
flask_jsonpify
couchdb
jsonpickle
fuzzywuzzy
pytest
